Major rewrite for better modal view, now with zoom function
The modal view now always uses 3 <div> elemnts for the images:
the current, previous and next image.

Touch gestures to swipe to the previous and next image are now
implemented using touch events and CSS transitions instead of
scroll events.  This scales better with large galleries and
fixes several issues with the scroll behaviour.